AI: Merge SrsHybridServer to SrsServer.
