+++ trunk/src/app/srs_app_rtc_conn.hpp
@@ -10,7 +10,6 @@
 #include <srs_app_async_call.hpp>
 #include <srs_app_conn.hpp>
 #include <srs_app_hourglass.hpp>
-#include <srs_app_hybrid.hpp>
 #include <srs_app_listener.hpp>
 #include <srs_app_reload.hpp>
 #include <srs_app_rtc_dtls.hpp>
