+++ trunk/src/app/srs_app_rtc_server.hpp
@@ -11,14 +11,13 @@
 
 #include <srs_app_async_call.hpp>
 #include <srs_app_hourglass.hpp>
-#include <srs_app_hybrid.hpp>
 #include <srs_app_listener.hpp>
 #include <srs_app_reload.hpp>
 #include <srs_app_rtc_sdp.hpp>
 #include <srs_app_st.hpp>
 
-#include <string>
 #include <set>
+#include <string>
 
 class SrsRtcServer;
 class SrsHourGlass;
